uint8 overflow corrected

+++ utils.h
@@ -12,7 +12,7 @@ uint32_t strTou32(const char *s);
 std::string u32str(uint32_t n);
 std::string u16str(uint16_t n);
 std::string u8str(uint8_t n);
-void coef(const std::string &S, const uint8_t w, uint8_t *dest, const uint8_t num);
+void coef(const std::string &S, const uint8_t w, uint8_t *dest, const uint16_t num);
 std::string cksm(const std::string &S, const uint8_t w, const uint8_t n, const uint16_t ls);
 
 //void cout_hex(const uint8_t *x, const size_t n);
